Move linux operations to their own package
So that I can use them from ssh auth-keys
